+++ src/mechanics/physics/Planet.java
@@ -7,11 +7,11 @@ import tensor.DVector;
 
 public class Planet extends Body {
 	
-	public Planet(DVector pos, float mass, float radius, Color color) {
+	public Planet(DVector pos, double mass, float radius, Color color) {
 		super(pos, mass, new Sphere(radius), color);
 	}
 	
-	public Planet(DVector pos, DVector vi, float mass, float radius, Color color) {
+	public Planet(DVector pos, DVector vi, double mass, float radius, Color color) {
 		super(pos, vi, mass, new Sphere(radius), color);
 	}
